refactor(last_common_ancestor): move LCA to separate module
