+++ src/tralda/datastructures/__init__.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from tralda.datastructures.tree import Tree as Tree
 from tralda.datastructures.tree import TreeNode as TreeNode
-from tralda.datastructures.tree import LCA as LCA
+from tralda.datastructures.last_common_ancestor import LCA as LCA
 from tralda.datastructures.linked import LinkedList as LinkedList
 from tralda.datastructures.linked import LinkedListNode as LinkedListNode
 from tralda.datastructures.doubly_linked import DLList as DLList
